requirements.txt
Django==2.1.2
django-crispy-forms==1.7.2
djangorestframework==3.8.2
mysqlclient==1.3.13
pytz==2018.5